backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-generativeai>=0.8.0
openai>=1.50.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.0
faiss-cpu>=1.13.0
pdfplumber>=0.11.0
pandas>=2.2.0
openpyxl>=3.1.2
reportlab>=4.2.0
python-multipart>=0.0.9
pydantic>=2.9.0
numpy>=1.26.0
requests>=2.32.0
